+++ ff_standard_lib/src/helpers/decimal_calculators.rs
@@ -14,7 +14,7 @@ pub fn divide_f64(dividend: f64, divisor: f64) -> f64 {
     result.to_f64().expect("Error converting result to f64")
 }
 
-pub fn round_to_decimals(value: f64, decimals: u32) -> f64 {
+pub fn round_to_decimals(value: f64, decimals: u64) -> f64 {
     let factor = 10f64.powi(decimals as i32);
     (value * factor).round() / factor
 }
